Change entrypoint of watch script into js
